UI, OrderingTable: remove interface 'Row'
